Removes "file presente in multiple build targets" warning
