+++ Cargo.toml
@@ -23,4 +23,4 @@ path = "src/main.rs"
 
 [[bin]]
 name = "jwtinfo"
-path = "src/main.rs"
+path = "src/cli.rs"
